refactor : 서비스 인터페이스, 구현체 구현, entity 접근 범위 제한
